Fix time interval
Fixes a bug in the time interval code that would cause 'emit's to flood in bursts. Because it was adding on the time interval to `nextUpdate` each time instead of adding it onto the current time, `nextUpdate` would slowly get more and more 'in the past' causing all future emits to flood until `nextUpdate` is incremented into the future again.